\name{RGEOS Experimental Functions}
\alias{poly_findInBoxGEOS}

\title{Experimental Functions}
\description{Functions still under development}
\usage{
    poly_findInBoxGEOS(spl, as_points=TRUE)
}

\arguments{
  \item{spl}{Object that inherits from the \code{SpatialPolygons} class}
  \item{as_points}{Logical value indicating if the polygon should be treated as points}
}


\author{Roger Bivand & Colin Rundel}

\keyword{spatial}

